django-template/requirements.txt
Django>=4.2,<5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
django-storages[s3]>=1.14.0
boto3>=1.34.0
whitenoise>=6.6.0
django-environ>=0.11.0
django-health-check>=3.18.0
